requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
docopt==0.6.2
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
pipreqs==0.4.13
protobuf==3.20.3
Pygments==2.15.1
PySocks==1.7.1
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.4
sqlparse==0.4.4
urllib3==1.26.15
yarg==0.1.9
